birdtag-system/birdtag-system-6.6.3/thumbnail_birddetectioin/birddetection_sagemaker/requirements.txt
roboflow 
ultralytics 
supervision
opencv-python
matplotlib
numpy
requests
flask
torch
pillow
